+++ src/EventInfo/EventInfo.js
